component: created the bosawas-ui button
This is the first major UI component to be added to Bosawas aside from
the text. It has the capability to add user inputted colors while
automatically adjusting the foreground. It can open links in new tabs.
It also just flat out looks cool.

Finally, this component also comes bundled with the inclusion of
utilities such as foreground prediction based on a variety of background
color inputs, as well as the inclusion of user events while testing.

- ref: https://stackoverflow.com/questions/36721830/convert-hsl-to-rgb-and-hex
- ref: https://stackoverflow.com/questions/11867545/change-text-color-based-on-brightness-of-the-covered-background-area